[improve][client] PIP-478: move the binary auth driver out of the public API module
AsyncAuthenticationDriver lived in pulsar-client-api as a "stable internal" marker, because
it existed for a carve-out: a v4 Authentication could additionally implement it and
ClientCnx would then drive that plugin asynchronously instead of synchronously. Since the
client drives the v5 model natively there is no such choice — the client resolves the
authentication it drives and wraps it itself, and nothing outside pulsar-client implements
or observes the type.

Move it to org.apache.pulsar.client.impl.auth.v5 and name it for what it now is,
BinaryAuthenticationDriver. That removes an internal type from the API module's surface, and
its javadoc no longer has to describe a synchronous alternative that no longer exists.

Also make the client's own blocking auth executor queue rather than reject. It was a
SynchronousQueue with the default abort policy, so the seventeenth concurrent credential call
failed outright — and every caller is a connection attempt or an authenticated request, so a
reconnect storm against a slow identity provider produced failed connections rather than slow
ones. This is the same fix the shared fallback pool already received, and it is what
pip-478.md already claims the executor does.
