+++ pulsar-client-api/src/main/java/org/apache/pulsar/client/api/internal/package-info.java
@@ -19,6 +19,6 @@
 
 /**
  * Stable-internal client API types observed by the framework but not intended for application
- * implementation. See {@link org.apache.pulsar.client.api.internal.AsyncAuthenticationDriver}.
+ * implementation. See {@link org.apache.pulsar.client.api.internal.BinaryAuthenticationDriver}.
  */
 package org.apache.pulsar.client.api.internal;
